Views refactoring
