+++ src/main/java/module-info.java
@@ -3,9 +3,13 @@ module org.openjfx.api2semestre {
     requires transitive javafx.fxml;
     requires transitive java.sql;
 
+    opens org.openjfx.api2semestre.view_controllers.templates to javafx.fxml;
+    opens org.openjfx.api2semestre.view_controllers.popups to javafx.fxml;
     opens org.openjfx.api2semestre.view_controllers.login to javafx.fxml;
     opens org.openjfx.api2semestre.view_controllers to javafx.fxml;
+
     opens org.openjfx.api2semestre.custom_tags to javafx.fxml;
     opens org.openjfx.api2semestre.view_utils to javafx.base;
+    
     exports org.openjfx.api2semestre;
 }
