+++ src/main/java/org/openjfx/API2Semestre/view_controllers/login/LoginAdm.java
@@ -34,8 +34,8 @@ public class LoginAdm {
             ))) {
                 App.changeView(ProvisoryLogin.handleViews());
             }
-        } catch (Exception e) {
-
+        } catch (Exception ex) {
+            ex.printStackTrace();
         }
     }
