+++ src/main/java/org/openjfx/API2Semestre/view_controllers/login/LoginCol.java
@@ -35,8 +35,8 @@ public class LoginCol {
             ))) {
                 App.changeView(ProvisoryLogin.handleViews());
             }
-        } catch (Exception e) {
-
+        } catch (Exception ex) {
+            ex.printStackTrace();
         }
     }
