+++ src/main/java/org/openjfx/API2Semestre/view_controllers/login/LoginGes.java
@@ -34,8 +34,8 @@ public class LoginGes {
             ))) {
                 App.changeView(ProvisoryLogin.handleViews());
             }
-        } catch (Exception e) {
-
+        } catch (Exception ex) {
+            ex.printStackTrace();
         }
     }
