+++ src/main/java/org/openjfx/API2Semestre/view_utils/PrettyTableCell.java
@@ -6,7 +6,7 @@ import javafx.scene.control.TableCell;
 
 public class PrettyTableCell<S, T> extends TableCell<S, T> {
     // private final TableColumn<S, T> column;
-    private PrettyTableCellInstruction<S, T>[] instructions = new PrettyTableCellInstruction[0];
+    private PrettyTableCellInstruction<S, T>[] instructions;
 
     // public PrettyTableCell(TableColumn<S, T> column, PrettyTableCellInstruction<S, T>[] instructions) {
     public PrettyTableCell(PrettyTableCellInstruction<S, T>[] instructions) {
@@ -18,13 +18,14 @@ public class PrettyTableCell<S, T> extends TableCell<S, T> {
     protected void updateItem(T item, boolean empty) {
         super.updateItem(item, empty);
         setAlignment(Pos.CENTER);
-        for (var instruction : instructions) {
-            if (instruction.apply(this, item, empty)) break;
-        }
         if (empty || item == null) {
             setText(null);
-        } else {
-            setText(item.toString());
+            return;
+        }
+        setText(item.toString());
+        if (instructions == null) return;
+        for (var instruction : instructions) {
+            if (instruction.apply(this, item, empty)) break;
         }
     }
 }
