+++ src/main/resources/org/openjfx/api2semestre/login/provisory_ges.fxml
@@ -1,0 +1,83 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<VBox alignment="CENTER" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="800.0" spacing="8.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.openjfx.api2semestre.view_controllers.login.LoginGes">
+   <children>
+      <ImageView fitHeight="61.0" fitWidth="131.0" pickOnBounds="true" preserveRatio="true">
+         <image>
+            <Image url="@imagens/logoKhali.png" />
+         </image>
+      </ImageView>
+      <Label alignment="CENTER" contentDisplay="CENTER" prefHeight="108.0" prefWidth="736.0" text="Identificação de Gestor">
+         <font>
+            <Font name="Franklin Gothic Medium" size="32.0" />
+         </font>
+      </Label>
+      <VBox spacing="8.0">
+         <children>
+            <HBox alignment="CENTER">
+               <children>
+                  <Label alignment="CENTER" contentDisplay="CENTER" prefHeight="27.0" prefWidth="169.0" text="Insira seu nome:">
+                     <font>
+                        <Font size="18.0" />
+                     </font>
+                  </Label>
+                  <TextField fx:id="tf_name" maxWidth="400.0" prefHeight="25.0" prefWidth="158.0" />
+               </children>
+            </HBox>
+            <VBox>
+               <children>
+                  <Label alignment="CENTER" contentDisplay="CENTER" prefHeight="56.0" prefWidth="718.0" text="Insira os Centros de Resultado que gerencia: (separados por vírgula)" textAlignment="CENTER" wrapText="true">
+                     <font>
+                        <Font size="18.0" />
+                     </font>
+                  </Label>
+                  <TextField fx:id="tf_manages" prefHeight="80.0" prefWidth="250.0" />
+               </children>
+               <padding>
+                  <Insets bottom="8.0" left="8.0" right="8.0" top="8.0" />
+               </padding>
+            </VBox>
+            <VBox layoutX="18.0" layoutY="53.0">
+               <children>
+                  <Label alignment="CENTER" contentDisplay="CENTER" prefHeight="56.0" prefWidth="718.0" text="Insira os Centros de Resultado que pertence: (separados por vírgula)" textAlignment="CENTER" wrapText="true">
+                     <font>
+                        <Font size="18.0" />
+                     </font>
+                  </Label>
+                  <TextField fx:id="tf_member" prefHeight="80.0" prefWidth="250.0" />
+               </children>
+               <padding>
+                  <Insets bottom="8.0" left="8.0" right="8.0" top="8.0" />
+               </padding>
+            </VBox>
+         </children>
+         <VBox.margin>
+            <Insets left="16.0" right="16.0" top="16.0" />
+         </VBox.margin>
+         <padding>
+            <Insets bottom="8.0" left="8.0" right="8.0" top="8.0" />
+         </padding>
+      </VBox>
+      <Button onAction="#login" alignment="CENTER" contentDisplay="CENTER" mnemonicParsing="false" text="Entrar">
+         <font>
+            <Font size="18.0" />
+         </font>
+         <padding>
+            <Insets bottom="8.0" left="64.0" right="64.0" top="8.0" />
+         </padding>
+      </Button>
+   </children>
+   <padding>
+      <Insets bottom="32.0" left="32.0" right="32.0" top="32.0" />
+   </padding>
+</VBox>
