+++ src/main/resources/org/openjfx/api2semestre/templates/viewButtonTemplate.fxml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.text.Font?>
+
+
+<Button 
+   fx:id="button"
+   mnemonicParsing="false" 
+   onAction="#enterView" 
+   prefHeight="25.0" 
+   prefWidth="216.0" 
+   text="Button" 
+   xmlns="http://javafx.com/javafx/19" 
+   xmlns:fx="http://javafx.com/fxml/1" 
+   fx:controller="org.openjfx.api2semestre.view_controllers.templates.ViewButtonController" 
+>
+   <font>
+      <Font size="16.0" />
+   </font>
+</Button>
