--- src/main/resources/org/openjfx/api2semestre/viewButtonTemplate.fxml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.text.Font?>
-
-
-<Button mnemonicParsing="false" onAction="#enterView" prefHeight="25.0" prefWidth="216.0" text="Button" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="BaseController">
-   <font>
-      <Font size="16.0" />
-   </font>
-</Button>
